requirements.txt
ffmpegio-core
simpleaudio
pyaudio